Cambio tipos int por bigint

+++ inicioBBDD.sql
@@ -215,9 +215,9 @@ DROP TABLE IF EXISTS `permission_user`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `permission_user` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `permission_id` int(10) unsigned NOT NULL,
-  `user_id` int(10) unsigned NOT NULL,
+  `id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
+  `permission_id` bigint(10) unsigned NOT NULL,
+  `user_id` bigint(10) unsigned NOT NULL,
   `created_at` timestamp NULL DEFAULT NULL,
   `updated_at` timestamp NULL DEFAULT NULL,
   PRIMARY KEY (`id`),
